+++ core/src/main/java/de/brainstormsoftworks/taloonerrl/core/engine/InputSystem.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     David Becker - initial API and implementation
  ******************************************************************************/
-package de.brainstormsoftworks.taloonerrl.system;
+package de.brainstormsoftworks.taloonerrl.core.engine;
 
 import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.InputAdapter;
